requirements.txt
# Core dependencies for Physics-Informed Sonar Detection System

# Deep learning framework
torch>=1.12.0
torchvision>=0.13.0

# Data handling and numerical computing
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
pillow>=8.3.0

# Configuration and serialization
pyyaml>=6.0
dataclasses-json>=0.5.7

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Data processing and analysis
pandas>=1.3.0
scikit-learn>=1.0.0

# Progress bars and utilities
tqdm>=4.62.0

# Testing (optional)
pytest>=6.2.0
pytest-cov>=2.12.0

# Development tools (optional)
black>=21.0.0
flake8>=3.9.0
mypy>=0.910